Inventory-Management-System-Backend/requirements.txt
# fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
